grpc_service/pyproject.toml
[tool.poetry]
name = "grpc-image-service"
version = "0.1.0"
description = ""
authors = ["Heyul <heyul.chavez@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
opencv-python-headless = "^4.9.0.80"
numpy = "^1.24.0"
pydantic = "^1.10.13"
tzdata = "^2024.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
